# Remove allocation when inserting into sublist