@nuss-justin 2016-04-06
Oh. That failure was surprising. I cann't reproduce that race on my machine, but it seems completely unrelated to my change.

From the stack traces, it seems like the server is shutdown, while still processing messages.