@nuss-justin 2016-04-06
Okay, I can't currently reproduce on my linux laptop. On my (much slower) windows machine I could reproduce it at least two times.  

@derekcollison I currently think the failure is not related to the change. Have you seen this error before or do you have some advice on how to debug this?